requirements.txt
torch>=1.9.0
torchaudio>=0.9.0
numpy>=1.21.0
matplotlib>=3.3.0
tqdm>=4.60.0
datasets>=2.0.0
transformers>=4.20.0
librosa>=0.9.0
soundfile>=0.10.0
torchcodec>=0.1.0
openai-whisper>=20231117
pennylane>=0.30.0
tiktoken>=0.5.0
numba>=0.56.0
jiwer>=4.0.0
editdistance>=0.8.0
